--- a/crates/bevy_ecs/src/bundle.rs
+++ b/crates/bevy_ecs/src/bundle.rs
@@ -15,6 +15,7 @@ use crate::{
         RequiredComponents, StorageType, Tick,
     },
     entity::{Entities, Entity, EntityLocation},
+    inheritance::InheritedComponents,
     observer::Observers,
     prelude::World,
     query::DebugCheckedUnwrap,
@@ -733,6 +734,8 @@ impl BundleInfo {
         storages: &mut Storages,
         components: &Components,
         observers: &Observers,
+        entities: &Entities,
+        inherited_components: &mut InheritedComponents,
         archetype_id: ArchetypeId,
     ) -> ArchetypeId {
         if let Some(archetype_after_insert_id) = archetypes[archetype_id]
@@ -836,6 +839,12 @@ impl BundleInfo {
                 table_components,
                 sparse_set_components,
             );
+            inherited_components.init_inherited_components(
+                entities,
+                components,
+                archetypes,
+                new_archetype_id,
+            );
             // Add an edge from the old archetype to the new archetype.
             archetypes[archetype_id]
                 .edges_mut()
@@ -872,6 +881,8 @@ impl BundleInfo {
         storages: &mut Storages,
         components: &Components,
         observers: &Observers,
+        entities: &Entities,
+        inherited_components: &mut InheritedComponents,
         archetype_id: ArchetypeId,
         intersection: bool,
     ) -> Option<ArchetypeId> {
@@ -946,6 +957,12 @@ impl BundleInfo {
                 next_table_id,
                 next_table_components,
                 next_sparse_set_components,
+            );
+            inherited_components.init_inherited_components(
+                entities,
+                components,
+                archetypes,
+                new_archetype_id,
             );
             Some(new_archetype_id)
         };
@@ -1028,6 +1045,8 @@ impl<'w> BundleInserter<'w> {
             &mut world.storages,
             &world.components,
             &world.observers,
+            &world.entities,
+            &mut world.inherited_components,
             archetype_id,
         );
         if new_archetype_id == archetype_id {
@@ -1165,6 +1184,21 @@ impl<'w> BundleInserter<'w> {
                 (archetype, location, after_effect)
             }
             ArchetypeMoveType::NewArchetypeSameTable { new_archetype } => {
+                {
+                    let new_archetype_id = new_archetype.as_ref().id();
+                    let old_archetype_id = self.archetype.as_ref().id();
+                    let world = self.world.world_mut();
+                    world
+                        .inherited_components
+                        .update_inherited_archetypes::<false>(
+                            &mut world.archetypes,
+                            &world.components,
+                            old_archetype_id,
+                            new_archetype_id,
+                            entity,
+                        );
+                }
+
                 let new_archetype = new_archetype.as_mut();
 
                 // SAFETY: Mutable references do not alias and will be dropped after this block
@@ -1209,6 +1243,21 @@ impl<'w> BundleInserter<'w> {
                 new_archetype,
                 new_table,
             } => {
+                {
+                    let new_archetype_id = new_archetype.as_ref().id();
+                    let old_archetype_id = self.archetype.as_ref().id();
+                    let world = self.world.world_mut();
+                    world
+                        .inherited_components
+                        .update_inherited_archetypes::<true>(
+                            &mut world.archetypes,
+                            &world.components,
+                            old_archetype_id,
+                            new_archetype_id,
+                            entity,
+                        );
+                }
+
                 let new_table = new_table.as_mut();
                 let new_archetype = new_archetype.as_mut();
 
@@ -1399,6 +1448,8 @@ impl<'w> BundleSpawner<'w> {
             &mut world.storages,
             &world.components,
             &world.observers,
+            &world.entities,
+            &mut world.inherited_components,
             ArchetypeId::EMPTY,
         );
         let archetype = &mut world.archetypes[new_archetype_id];