… crates/bevy_ecs/Cargo.toml
… crates/bevy_ecs/src/query/filter.rs
… crates/bevy_ecs/src/query/iter.rs
… crates/bevy_ecs/src/query/mod.rs
… crates/bevy_ecs/src/query/state.rs
… crates/bevy_ecs/src/query/world_query.rs
… crates/bevy_ecs/src/reflect/component.rs
… crates/bevy_ecs/src/storage/sparse_set.rs
… crates/bevy_ecs/src/system/commands/mod.rs
… crates/bevy_ecs/src/system/function_system.rs
… crates/bevy_ecs/src/system/system_param.rs
… crates/bevy_ecs/src/world/component_constants.rs
… crates/bevy_ecs/src/world/deferred_world.rs
… crates/bevy_ecs/src/world/entity_ref.rs
… crates/bevy_ecs/src/world/mod.rs
… crates/bevy_ecs/src/world/unsafe_world_cell.rs
… crates/bevy_gizmos/src/gizmos.rs
… crates/bevy_render/src/sync_world.rs
… crates/bevy_text/src/text_access.rs
… crates/bevy_transform/src/systems.rs
… crates/bevy_ui/src/widget/text.rs
… crates/bevy_winit/src/state.rs