# Objective
This PR adds an implementation of inherited components and [entity prefabs](https://www.flecs.dev/flecs/md_docs_2PrefabsManual.html). This is useful for creating dynamic entity templates and allows component sharing, which reduces memory pressure when there are many entities with similar components. Some of the prime example of where this is useful are entity-per-tile tilemaps and ecs-based physics implementations.

## Solution
Any entity can add an `InheritFrom` component to opt-into inheriting components from the target entity.
```rs
let base = world.spawn(ComponentA).id();
let inherited = world.spawn(InheritFrom(base)).id();

let base_entity = world.entity(base);
let inherited_entity = world.entity(inherited);

let base_component_a = base_entity.get::<ComponentA>();
assert_eq!(inherited_entity.get::<ComponentA>(), base_component_a);
```

Any changes to base entity will affect all entities inheriting from it as well. Inherited components are stored only on the base entity, so all entities inheriting a component from the base entity share the same component data.

Getting a shared component is supported only for non-mutable references for now, trying to get a mutable reference to a shared component will result in the same behavior as if requested component doesn't exist on the entity.

## Implementation
The implementation is split into several commits, each containing a logical chunk to make reviewing it easier.

There's a new field on world, `InheritedComponents`, that has all the data needed to resolve inherited components for archetypes/tables. If archetype/table has an `InheritFrom` component, it is opted-into resolving inherited components during iteration.

Most of the functionality is active only on tables/archetypes that have inherited components, so this shouldn't impact performance of normal queries too much. However, iteration performance benchmarks seem to suffer anyway and need more investigation.

One other problem is with systems that have queries working with both inherited components and base components - this system should conflict, but it doesn't right now:
```rs
fn query_system(
    base_q: Query<&mut CompA, With<Inherited>>,
    inherited_q: Query<&CompA, Without<Inherited>>,
) {
}
```
If anyone has any ideas about how to fix this problem and the iteration performance, it'd solve most of the blockers for this PR (from my perspective).

## Testing
Added tests for all the basic functionality.
Not all test pass yet, and there might be more cases that need testing.

## Future work
This implementation implements only the basic parts of component inheritance, some important stuff that's still missing:
- Make required components' requirement fulfilled if inherited entity contains the component
- Support mutable references for inherited components
- Support filtering which components get inherited on the base entity
- Support filtering which components get inherited on the inherited entity
- Allow marking components as not inheritable