@alice-i-cecile 2025-04-08
I did some complimentary work over in #17769, working on the "create a library of entities to draw from" pattern using disabling components.