@eugineerd 2025-04-14
Been stuck in optimization hell staring at assembly code and llvm ir while implementing shared mutable components, but I've finally managed to implement the basic version of this. Here's the results of my struggles (this is just the iteration benchmarks):
<details><summary>Iteration benchmarks</summary>

| **Benchmark**                              | **Main (ns/µs)** | **PR (ns/µs)**   | **Change from Second**           |
| ------------------------------------------ | ---------------- | ---------------- | -------------------------------- |
| iter_fragmented/base                       | 314.00-318.83 ns | 327.05-332.22 ns | +2.99% to +4.60% (regressed)     |
| iter_fragmented/wide                       | 4.8318-4.8808 µs | 4.8904-4.9244 µs | +0.82% to +2.56% (noise)         |
| iter_fragmented/foreach                    | 150.64-159.99 ns | 141.75-150.94 ns | -10.58% to -3.87% (improved)     |
| iter_fragmented/foreach_wide               | 5.9477-6.0151 µs | 18.166-18.228 µs | +204.39% to +207.85% (regressed) |
| iter_fragmented_sparse/base                | 7.3423-7.8950 ns | 7.8041-7.9021 ns | +0.13% to +6.02% (noise)         |
| iter_fragmented_sparse/wide                | 48.128-48.368 ns | 80.809-81.130 ns | +67.26% to +68.52% (regressed)   |
| iter_fragmented_sparse/foreach             | 7.3979-8.3397 ns | 7.3064-7.4126 ns | -9.68% to -1.02% (improved)      |
| iter_fragmented_sparse/foreach_wide        | 75.540-75.974 ns | 212.62-213.90 ns | +179.89% to +183.44% (regressed) |
| iter_simple/base                           | 6.8433-6.8954 µs | 6.8330-6.8513 µs | -0.44% to +0.95% (no change)     |
| iter_simple/wide                           | 41.296-41.581 µs | 65.385-65.533 µs | +57.34% to +58.83% (regressed)   |
| iter_simple/system                         | 6.8414-6.8629 µs | 6.8550-6.8718 µs | -0.09% to +0.36% (no change)     |
| iter_simple/sparse_set                     | 21.107-21.186 µs | 21.164-21.315 µs | +0.26% to +0.87% (noise)         |
| iter_simple/wide_sparse_set                | 89.770-90.504 µs | 102.47-102.87 µs | +13.92% to +14.81% (regressed)   |
| iter_simple/foreach                        | 6.8061-6.8267 µs | 6.7805-6.8012 µs | -0.52% to -0.01% (no change)     |
| iter_simple/foreach_wide                   | 31.290-31.402 µs | 46.171-46.293 µs | +47.20% to +48.19% (regressed)   |
| iter_simple/foreach_sparse_set             | 20.834-20.938 µs | 20.650-20.726 µs | -1.55% to -1.03% (improved)      |
| iter_simple/foreach_wide_sparse_set        | 96.508-96.770 µs | 116.87-117.19 µs | +20.51% to +21.13% (regressed)   |
| iter_simple/foreach_hybrid                 | 8.8300-8.8746 µs | 9.0273-9.0633 µs | +1.91% to +2.43% (regressed)     |
| par_iter_simple/with_0_fragment            | 36.391-36.702 µs | 36.334-36.761 µs | -0.42% to +2.29% (no change)     |
| par_iter_simple/with_10_fragment           | 36.630-36.877 µs | 36.759-37.115 µs | -0.54% to +2.10% (no change)     |
| par_iter_simple/with_100_fragment          | 37.188-37.656 µs | 37.434-38.180 µs | -1.07% to +1.94% (no change)     |
| par_iter_simple/with_1000_fragment         | 46.469-47.187 µs | 47.155-47.807 µs | -0.43% to +2.81% (no change)     |
| par_iter_simple/hybrid                     | 74.036-74.795 µs | 74.586-75.909 µs | -1.08% to +2.62% (no change)     |
| iter_fragmented(4096)_empty/foreach_table  | 3.2693-3.3137 µs | 2.8646-2.8891 µs | -12.61% to -11.11% (improved)    |
| iter_fragmented(4096)_empty/foreach_sparse | 10.008-10.078 µs | 11.040-11.087 µs | +9.76% to +10.70% (regressed)    |

</details>

The results are... not too bad, I think. There is a clear regression for wide queries, but smaller queries seem to perform just fine. It _might_ be possible to optimize wide queries, but I doubt it. Looking at the generated assembly, there is a branch on each access (`test` followed by `jnz`), so I believe this is just hitting the limitations of the branch predictor at this point. All of the benchmarks are really tight loops, so I suspect the impact on real-world code that has much more going on will be somewhat different. 
Some of the benchmarks that seem to have improved are probably due to the additional `#[inline]`'s I had to sprinkle throughout the fetch implementation.

This implementation uses component_id + table_id/archetype_id as a key for the staged mutated components hashmap that stores the cloned components on `deref_mut`, so in `QuerySet` the conflicting queries with the same shared mutable components will see the changes introduced by each other. Then, on `world.flush`, the staged components are inserted on the target entity.

I need to clean up the implementation a bit before I push the changes here, but overall I think there might be a path forward for this feature (if we're ok with taking the hit for wide queries).